Organized classes into packages, deleted unused classes, and added another check for .txt file extension when using file upload.
