+++ TAB2MXL/src/main/java/gui_helper/ColorDef.java
@@ -1,4 +1,4 @@
-package tab2mxl;
+package gui_helper;
 
 
 /**
