+++ TAB2MXL/src/main/java/gui_popups/ClearPopUp.java
@@ -1,8 +1,7 @@
-package tab2mxl;
+package gui_popups;
 
 import java.awt.Color;
 import java.awt.Component;
-import java.awt.Desktop;
 import java.awt.Dimension;
 import java.awt.FlowLayout;
 import java.awt.Font;
@@ -10,8 +9,6 @@ import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 import java.awt.event.MouseAdapter;
 import java.awt.event.MouseEvent;
-import java.io.File;
-import java.io.IOException;
 import java.awt.Point;
 
 import javax.swing.BorderFactory;
@@ -23,20 +20,25 @@ import javax.swing.Popup;
 import javax.swing.PopupFactory;
 import javax.swing.border.Border;
 
-public class SuccessPopUp extends PopupFactory implements ActionListener{
+import gui_panels.MyFrame;
+import tab2mxl.Main;
+
+public class ClearPopUp extends PopupFactory implements ActionListener{
 
 	Popup popup;
 	
-	JButton okButton;
-	JButton openButton;
+	JButton yesButton;
+	JButton noButton;
 	
-	MyFrame myFrame;
-	String path;
-	
-	SuccessPopUp (Component owner, String path){
+	String input;
+	MyFrame myFrame;	
+		
+	public ClearPopUp (Component owner, String input, String message){
 				
+		
+		
+		this.input = input;
 		myFrame = (MyFrame) owner;
-		this.path = path;
 		
         JPanel panel = new JPanel(); 
         panel.setPreferredSize(new Dimension(300, 100));
@@ -47,63 +49,61 @@ public class SuccessPopUp extends PopupFactory implements ActionListener{
         
         JPanel titlepanel = new JPanel();
         titlepanel.setLayout(new FlowLayout());
-        JLabel title = new JLabel("Conversion Was Successful");
-        title.setForeground(new Color(110,199,56));
+        JLabel title = new JLabel(message);
+        title.setForeground(new Color(232, 32, 21));
         title.setFont(new Font(title.getName(), Font.BOLD, 14));
         Border titlePadding = BorderFactory.createEmptyBorder(15, 10, 0, 10);
         title.setBorder(titlePadding);
         titlepanel.add(title);
         
         JPanel inputPanel = new JPanel();
-        inputPanel.setLayout(new FlowLayout());
+        inputPanel.setLayout(new FlowLayout());     
         
-        //Ok button
-        okButton = new JButton("Ok");        
-        okButton.setBorderPainted(false);
-        okButton.setBackground(new Color(33,150,243));
-        okButton.setForeground(new Color(224,224,224));
-        okButton.setFocusable(false);
-        okButton.addActionListener(this);
-        okButton.setOpaque(true);
-        okButton.setBorderPainted(false);
+        yesButton = new JButton("Yes");
+        yesButton.setBorderPainted(false);
+        yesButton.setBackground(new Color(33,150,243));
+        yesButton.setForeground(new Color(224,224,224));
+        yesButton.setFocusable(false);
+        yesButton.addActionListener(this);
+        yesButton.setOpaque(true);
+        yesButton.setBorderPainted(false);
         
         //Button hover effects
-        okButton.addMouseListener(new java.awt.event.MouseAdapter() {
+        yesButton.addMouseListener(new java.awt.event.MouseAdapter() {
     	    public void mouseEntered(java.awt.event.MouseEvent evt) {
-    	    	okButton.setBackground(new Color(60,160,243));
-    	    	okButton.setForeground(Color.black);
+    	    	yesButton.setBackground(new Color(60,160,243));
+    	    	yesButton.setForeground(Color.black);
     	    }
 
     	    public void mouseExited(java.awt.event.MouseEvent evt) {
-    	    	okButton.setBackground(new Color(33,150,243));
-    	    	okButton.setForeground(new Color(224,224,224));
+    	    	yesButton.setBackground(new Color(33,150,243));
+    	    	yesButton.setForeground(new Color(224,224,224));
     	    }
-    	});   
-        inputPanel.add(okButton);
+    	});        
+        inputPanel.add(yesButton);        
         
-        //Open button
-        openButton = new JButton("Open");        
-        openButton.setBorderPainted(false);
-        openButton.setBackground(new Color(33,150,243));
-        openButton.setForeground(new Color(224,224,224));
-        openButton.setFocusable(false);
-        openButton.addActionListener(this);
-        openButton.setOpaque(true);
-        openButton.setBorderPainted(false);
+        noButton = new JButton("No");
+        noButton.setBorderPainted(false);
+        noButton.setBackground(new Color(33,150,243));
+        noButton.setForeground(new Color(224,224,224));
+        noButton.setFocusable(false);
+        noButton.addActionListener(this);
+        noButton.setOpaque(true);
+        noButton.setBorderPainted(false);
         
         //Button hover effects
-        openButton.addMouseListener(new MouseAdapter() {
+        noButton.addMouseListener(new MouseAdapter() {
     	    public void mouseEntered(MouseEvent evt) {
-    	    	openButton.setBackground(new Color(60,160,243));
-    	    	openButton.setForeground(Color.black);
+    	    	noButton.setBackground(new Color(60,160,243));
+    	    	noButton.setForeground(Color.black);
     	    }
 
     	    public void mouseExited(MouseEvent evt) {
-    	    	openButton.setBackground(new Color(33,150,243));
-    	    	openButton.setForeground(new Color(224,224,224));
+    	    	noButton.setBackground(new Color(33,150,243));
+    	    	noButton.setForeground(new Color(224,224,224));
     	    }
-    	});   
-        inputPanel.add(openButton);
+    	});        
+        inputPanel.add(noButton);
         
         Border buttonPadding = BorderFactory.createEmptyBorder(0, 10, 0, 10);
         inputPanel.setBorder(buttonPadding);
@@ -117,26 +117,21 @@ public class SuccessPopUp extends PopupFactory implements ActionListener{
         popup = this.getPopup(myFrame, panel, pt.x + frameSize.width/2 - 150, pt.y + frameSize.height/2 - 50);
         popup.show();
         Main.isInPopUp = true;
-	}
+}
 
 	@Override
 	public void actionPerformed(ActionEvent e) {
-		if(e.getSource() == okButton)
+		if(e.getSource() == yesButton)
 		{
+			myFrame.textInputContentPanel.textField.setText(input);
 			popup.hide();
 			Main.isInPopUp = false;
 		}
-		else if(e.getSource() == openButton)
+		else
 		{
-			if (Desktop.isDesktopSupported()) {
-				try {
-					Desktop.getDesktop().open(new File(path).getParentFile());
-				} catch (IOException e1) {
-					e1.printStackTrace();
-				}
-			}
 			popup.hide();
 			Main.isInPopUp = false;
 		}
 	}	
+	
 }
