+++ TAB2MXL/src/main/java/tab2mxl/Main.java
@@ -2,6 +2,9 @@ package tab2mxl;
 
 import java.util.ArrayList;
 
+import gui_panels.MyFrame;
+import gui_popups.ClearPopUp;
+
 public class Main {
 	
 	public static MyFrame myFrame;
